+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Notable changes since the last release of the [SSI Kit](https://github.com/walt-
 
 ## [Unreleased]
 
+- Features
+  - Added GaiaxSelfDecription credential https://github.com/dNationCloud/waltid-ssikit/pull/1 thx to https://github.com/dNationCloud & https://github.com/matofeder
+
 ## [1.1.0] - 2021-11-25
 
 -   Features